+++ test/Bling.spec.js
@@ -1,10 +1,13 @@
 /* eslint-disable react/no-multi-comp */
-import React, {Component, PropTypes} from "react";
-import ReactTestUtils from "react-addons-test-utils";
+import React, {Component} from "react";
+import PropTypes from "prop-types";
+import ReactTestUtils from "react-dom/test-utils";
+import ShallowRenderer from "react-test-renderer/shallow";
 import Bling from "../src/Bling";
 import Events from "../src/Events";
 import {pubadsAPI, APIToCallBeforeServiceEnabled} from "../src/createManager";
 import {gptVersion, pubadsVersion} from "../src/utils/apiList";
+import {createManagerTest} from "../src/utils/createManagerTest";
 
 describe("Bling", () => {
     let googletag;
@@ -12,7 +15,7 @@ describe("Bling", () => {
 
     beforeEach(() => {
         Bling.configure({renderWhenViewable: false});
-        Bling.createTestManager();
+        Bling.testManager = createManagerTest();
         googletag = Bling._adManager.googletag;
     });
 
@@ -23,22 +26,27 @@ describe("Bling", () => {
     });
 
     it("throws when either slotSize or sizeMapping is missing", () => {
-        expect(() => {
-            ReactTestUtils.renderIntoDocument(
-                <Bling adUnitPath="/4595/nfl.test.open" />
-            );
-        }).to.throw("Either 'slotSize' or 'sizeMapping' prop needs to be set.");
+        const renderBling = () => {
+            const renderer = new ShallowRenderer();
+            renderer.render(<Bling adUnitPath="/4595/nfl.test.open" />);
+        };
+
+        expect(renderBling).to.throw(
+            "Either 'slotSize' or 'sizeMapping' prop needs to be set."
+        );
     });
 
     it("initially renders empty div with style", () => {
-        const renderer = ReactTestUtils.createRenderer();
-        renderer.render(<Bling adUnitPath="/4595/nfl.test.open" slotSize={[728, 90]} />);
+        const renderer = new ShallowRenderer();
+        renderer.render(
+            <Bling adUnitPath="/4595/nfl.test.open" slotSize={[728, 90]} />
+        );
         const result = renderer.getRenderOutput();
         expect(result.type).to.equal("div");
         expect(result.props.style).to.eql({width: 728, height: 90});
     });
 
-    it("returns gpt version", (done) => {
+    it("returns gpt version", done => {
         Bling.once(Events.READY, () => {
             expect(Bling.getGPTVersion()).to.equal(gptVersion);
             done();
@@ -49,7 +57,7 @@ describe("Bling", () => {
         );
     });
 
-    it("returns pubads version", (done) => {
+    it("returns pubads version", done => {
         Bling.once(Events.READY, () => {
             expect(Bling.getPubadsVersion()).to.equal(pubadsVersion);
             done();
@@ -60,12 +68,16 @@ describe("Bling", () => {
         );
     });
 
-    it("accepts syncCorrelator", (done) => {
-        const render = sinon.stub(Bling._adManager, "render", function () {
-            expect(this._syncCorrelator).to.be.true;
-            render.restore();
-            done();
-        });
+    it("accepts syncCorrelator", done => {
+        const render = sinon.stub(
+            Bling._adManager,
+            "render",
+            function syncCorrelator() {
+                expect(this._syncCorrelator).to.be.true;
+                render.restore();
+                done();
+            }
+        );
 
         Bling.syncCorrelator();
 
@@ -74,7 +86,7 @@ describe("Bling", () => {
         );
     });
 
-    it("accepts pubads API", (done) => {
+    it("accepts pubads API", done => {
         const apiStubs = {};
         pubadsAPI.forEach(method => {
             apiStubs[method] = sinon.stub(googletag.pubads(), method);
@@ -112,7 +124,7 @@ describe("Bling", () => {
         );
     });
 
-    it("can call pubads API multiple times", (done) => {
+    it("can call pubads API multiple times", done => {
         const spy = sinon.stub(googletag.pubads(), "setTargeting");
 
         Bling.once(Events.RENDER, () => {
@@ -131,7 +143,7 @@ describe("Bling", () => {
         );
     });
 
-    it("fires once event", (done) => {
+    it("fires once event", done => {
         const events = Object.keys(Events).map(key => Events[key]);
 
         function afterReady() {
@@ -154,7 +166,7 @@ describe("Bling", () => {
         );
     });
 
-    it("fires on event", (done) => {
+    it("fires on event", done => {
         const events = Object.keys(Events).map(key => Events[key]);
 
         function afterReady() {
@@ -178,7 +190,7 @@ describe("Bling", () => {
         );
     });
 
-    it("removes event", (done) => {
+    it("removes event", done => {
         const spy = sinon.spy();
         Bling.on(Events.RENDER, spy);
         Bling.removeListener(Events.RENDER, spy);
@@ -216,8 +228,28 @@ describe("Bling", () => {
         clear.restore();
     });
 
+    it("handles empty adSlot on clear", () => {
+        const instance = new Bling();
+        instance._adSlot = {};
+
+        expect(() => {
+            instance.clear();
+        }).to.not.throw("adSlot.getServices is not a function");
+    });
+
+    it("calls getServices on adSlot on clear", () => {
+        const instance = new Bling();
+        const adSlot = sinon.mock({getServices: () => {}});
+        adSlot.expects("getServices").once();
+        instance._adSlot = adSlot;
+        instance.clear();
+    });
+
     it("updates correlator", () => {
-        const updateCorrelator = sinon.stub(Bling._adManager, "updateCorrelator");
+        const updateCorrelator = sinon.stub(
+            Bling._adManager,
+            "updateCorrelator"
+        );
 
         Bling.updateCorrelator();
 
@@ -225,7 +257,7 @@ describe("Bling", () => {
         updateCorrelator.restore();
     });
 
-    it("reflects adUnitPath props to adSlot", (done) => {
+    it("reflects adUnitPath props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
             expect(adSlot.getAdUnitPath()).to.equal("/4595/nfl.test.open");
@@ -233,14 +265,11 @@ describe("Bling", () => {
         });
 
         const instance = ReactTestUtils.renderIntoDocument(
-            <Bling
-                adUnitPath="/4595/nfl.test.open"
-                slotSize={[300, 250]}
-            />
+            <Bling adUnitPath="/4595/nfl.test.open" slotSize={[300, 250]} />
         );
     });
 
-    it("reflects slotSize props to adSlot", (done) => {
+    it("reflects slotSize props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
             expect(adSlot.getSizes()).to.eql([300, 250]);
@@ -248,17 +277,18 @@ describe("Bling", () => {
         });
 
         const instance = ReactTestUtils.renderIntoDocument(
-            <Bling
-                adUnitPath="/4595/nfl.test.open"
-                slotSize={[300, 250]}
-            />
+            <Bling adUnitPath="/4595/nfl.test.open" slotSize={[300, 250]} />
         );
     });
 
-    it("reflects sizeMapping props to adSlot", (done) => {
+    it("reflects sizeMapping props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
-            expect(adSlot.getSizes()).to.eql([[[0, 0], [320, 50]], [[750, 200], [728, 90]], [[1050, 200], [1024, 120]]]);
+            expect(adSlot.getSizes()).to.eql([
+                [[0, 0], [320, 50]],
+                [[750, 200], [728, 90]],
+                [[1050, 200], [1024, 120]]
+            ]);
             done();
         });
 
@@ -274,7 +304,7 @@ describe("Bling", () => {
         );
     });
 
-    it("reflects targeting props to adSlot", (done) => {
+    it("reflects targeting props to adSlot", done => {
         const targeting = {t1: "v1", t2: [1, 2, 3]};
 
         Bling.once(Events.RENDER, () => {
@@ -294,7 +324,7 @@ describe("Bling", () => {
         );
     });
 
-    it("reflects collapseEmptyDiv props to adSlot", (done) => {
+    it("reflects collapseEmptyDiv props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             expect(ads[0].adSlot.getCollapseEmptyDiv()).to.be.true;
             expect(ads[1].adSlot.getCollapseEmptyDiv()).to.be.false;
@@ -304,11 +334,9 @@ describe("Bling", () => {
         class Wrapper extends Component {
             static propTypes = {
                 children: PropTypes.node
-            }
+            };
             render() {
-                return (
-                    <div>{this.props.children}</div>
-                );
+                return <div>{this.props.children}</div>;
             }
         }
 
@@ -326,10 +354,13 @@ describe("Bling", () => {
                 />
             </Wrapper>
         );
-        const ads = ReactTestUtils.scryRenderedComponentsWithType(instance, Bling);
+        const ads = ReactTestUtils.scryRenderedComponentsWithType(
+            instance,
+            Bling
+        );
     });
 
-    it("reflects attributes props to adSlot", (done) => {
+    it("reflects attributes props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
             expect(adSlot.get("attr1")).to.equal("val1");
@@ -347,7 +378,7 @@ describe("Bling", () => {
         );
     });
 
-    it("reflects categoryExclusion props to adSlot", (done) => {
+    it("reflects categoryExclusion props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             expect(ads[0].adSlot.getCategoryExclusions()).to.eql(["Airline"]);
             expect(ads[1].adSlot.getCategoryExclusions()).to.eql(["Airline"]);
@@ -357,11 +388,9 @@ describe("Bling", () => {
         class Wrapper extends Component {
             static propTypes = {
                 children: PropTypes.node
-            }
+            };
             render() {
-                return (
-                    <div>{this.props.children}</div>
-                );
+                return <div>{this.props.children}</div>;
             }
         }
 
@@ -379,10 +408,13 @@ describe("Bling", () => {
                 />
             </Wrapper>
         );
-        const ads = ReactTestUtils.scryRenderedComponentsWithType(instance, Bling);
+        const ads = ReactTestUtils.scryRenderedComponentsWithType(
+            instance,
+            Bling
+        );
     });
 
-    it("reflects clickUrl props to adSlot", (done) => {
+    it("reflects clickUrl props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
             expect(adSlot.getClickUrl()).to.equal("clickUrl");
@@ -398,11 +430,13 @@ describe("Bling", () => {
         );
     });
 
-    it("reflects companionAdService props to adSlot", (done) => {
+    it("reflects companionAdService props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
             const services = adSlot.getServices();
-            const companionAdService = services.filter(service => !!service.setRefreshUnfilledSlots)[0];
+            const companionAdService = services.filter(
+                service => !!service.setRefreshUnfilledSlots
+            )[0];
             expect(companionAdService._enableSyncLoading).to.be.true;
             expect(companionAdService._refreshUnfilledSlots).to.be.true;
             done();
@@ -411,13 +445,16 @@ describe("Bling", () => {
         const instance = ReactTestUtils.renderIntoDocument(
             <Bling
                 adUnitPath="/4595/nfl.test.open"
-                companionAdService={{enableSyncLoading: true, refreshUnfilledSlots: true}}
+                companionAdService={{
+                    enableSyncLoading: true,
+                    refreshUnfilledSlots: true
+                }}
                 slotSize={[300, 250]}
             />
         );
     });
 
-    it("reflects outOfPage props to adSlot", (done) => {
+    it("reflects outOfPage props to adSlot", done => {
         Bling.once(Events.RENDER, () => {
             const adSlot = instance.adSlot;
             expect(adSlot.getSizes()).to.eql([1, 1]);
@@ -425,14 +462,11 @@ describe("Bling", () => {
         });
 
         const instance = ReactTestUtils.renderIntoDocument(
-            <Bling
-                adUnitPath="/4595/nfl.test.open"
-                outOfPage={true}
-            />
+            <Bling adUnitPath="/4595/nfl.test.open" outOfPage={true} />
         );
     });
 
-    it("renders static ad", (done) => {
+    it("renders static ad", done => {
         const content = `<a href="www.mydestinationsite.com"><img src="www.mysite.com/img.png"></img></a>`;
 
         Bling.once(Events.RENDER, () => {
@@ -450,8 +484,12 @@ describe("Bling", () => {
         );
     });
 
-    it("does not render ad when renderWhenViewable prop is set to true and the component is not in viewport", (done) => {
-        const isInViewport = sinon.stub(Bling._adManager, "isInViewport", () => false);
+    it("does not render ad when renderWhenViewable prop is set to true and the component is not in viewport", done => {
+        const isInViewport = sinon.stub(
+            Bling._adManager,
+            "isInViewport",
+            () => false
+        );
 
         Bling.once(Events.SLOT_RENDER_ENDED, event => {
             expect(event.slot).to.equal(ads[1].adSlot);
@@ -466,11 +504,9 @@ describe("Bling", () => {
         class Wrapper extends Component {
             static propTypes = {
                 children: PropTypes.node
-            }
+            };
             render() {
-                return (
-                    <div>{this.props.children}</div>
-                );
+                return <div>{this.props.children}</div>;
             }
         }
 
@@ -489,11 +525,18 @@ describe("Bling", () => {
                 />
             </Wrapper>
         );
-        const ads = ReactTestUtils.scryRenderedComponentsWithType(instance, Bling);
+        const ads = ReactTestUtils.scryRenderedComponentsWithType(
+            instance,
+            Bling
+        );
     });
 
-    it("renders ad as soon as the component gets in the viewport", (done) => {
-        const isInViewport = sinon.stub(Bling._adManager, "isInViewport", () => false);
+    it("renders ad as soon as the component gets in the viewport", done => {
+        const isInViewport = sinon.stub(
+            Bling._adManager,
+            "isInViewport",
+            () => false
+        );
 
         Bling.once(Events.SLOT_RENDER_ENDED, event => {
             expect(event.slot).to.equal(ads[0].adSlot);
@@ -511,11 +554,9 @@ describe("Bling", () => {
         class Wrapper extends Component {
             static propTypes = {
                 children: PropTypes.node
-            }
+            };
             render() {
-                return (
-                    <div>{this.props.children}</div>
-                );
+                return <div>{this.props.children}</div>;
             }
         }
 
@@ -528,10 +569,13 @@ describe("Bling", () => {
                 />
             </Wrapper>
         );
-        const ads = ReactTestUtils.scryRenderedComponentsWithType(instance, Bling);
+        const ads = ReactTestUtils.scryRenderedComponentsWithType(
+            instance,
+            Bling
+        );
     });
 
-    it("refreshes ad when refreshable prop changes", (done) => {
+    it("refreshes ad when refreshable prop changes", done => {
         let count = 0;
 
         Bling.syncCorrelator();
@@ -539,7 +583,7 @@ describe("Bling", () => {
         class Wrapper extends Component {
             state = {
                 targeting: {prop: "val"}
-            }
+            };
             onSlotRenderEnded = event => {
                 if (count === 0) {
                     expect(event.slot.getTargeting("prop")).to.equal("val");
@@ -551,7 +595,7 @@ describe("Bling", () => {
                     expect(event.slot.getTargeting("prop")).to.equal("val2");
                     done();
                 }
-            }
+            };
             render() {
                 const {targeting} = this.state;
                 return (
@@ -565,12 +609,10 @@ describe("Bling", () => {
             }
         }
 
-        ReactTestUtils.renderIntoDocument(
-            <Wrapper />
-        );
+        ReactTestUtils.renderIntoDocument(<Wrapper />);
     });
 
-    it("refreshes ad when refreshableProps changes w/o sync correlator", (done) => {
+    it("refreshes ad when refreshableProps changes w/o sync correlator", done => {
         let count = 0;
 
         Bling.syncCorrelator(false);
@@ -578,7 +620,7 @@ describe("Bling", () => {
         class Wrapper extends Component {
             state = {
                 targeting: {prop: "val"}
-            }
+            };
             onSlotRenderEnded = event => {
                 if (count === 0) {
                     expect(event.slot.getTargeting("prop")).to.equal("val");
@@ -590,7 +632,7 @@ describe("Bling", () => {
                     expect(event.slot.getTargeting("prop")).to.equal("val2");
                     done();
                 }
-            }
+            };
             render() {
                 const {targeting} = this.state;
                 return (
@@ -604,12 +646,10 @@ describe("Bling", () => {
             }
         }
 
-        ReactTestUtils.renderIntoDocument(
-            <Wrapper />
-        );
+        ReactTestUtils.renderIntoDocument(<Wrapper />);
     });
 
-    it("re-renders ad when reRenderProps changes", (done) => {
+    it("re-renders ad when reRenderProps changes", done => {
         let count = 0;
 
         Bling.syncCorrelator();
@@ -617,19 +657,23 @@ describe("Bling", () => {
         class Wrapper extends Component {
             state = {
                 adUnitPath: "/4595/nfl.test.open"
-            }
+            };
             onSlotRenderEnded = event => {
                 if (count === 0) {
-                    expect(event.slot.getAdUnitPath()).to.equal("/4595/nfl.test.open");
+                    expect(event.slot.getAdUnitPath()).to.equal(
+                        "/4595/nfl.test.open"
+                    );
                     this.setState({
                         adUnitPath: "/4595/nfl.test.open/new"
                     });
                     count++;
                 } else {
-                    expect(event.slot.getAdUnitPath()).to.equal("/4595/nfl.test.open/new");
+                    expect(event.slot.getAdUnitPath()).to.equal(
+                        "/4595/nfl.test.open/new"
+                    );
                     done();
                 }
-            }
+            };
             render() {
                 const {adUnitPath} = this.state;
                 return (
@@ -642,17 +686,12 @@ describe("Bling", () => {
             }
         }
 
-        ReactTestUtils.renderIntoDocument(
-            <Wrapper />
-        );
+        ReactTestUtils.renderIntoDocument(<Wrapper />);
     });
 
     it("removes itself from registry when unmounted", () => {
         const instance = ReactTestUtils.renderIntoDocument(
-            <Bling
-                adUnitPath="/4595/nfl.test.open"
-                slotSize={[300, 250]}
-            />
+            <Bling adUnitPath="/4595/nfl.test.open" slotSize={[300, 250]} />
         );
         instance.componentWillUnmount();
         expect(Bling._adManager.getMountedInstances()).to.have.length(0);